pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-context"
version = "0.0.2"
authors = [
  { name="Vladimir Petrukhin", email="man4j@ya.ru" },
]
description = "Large Language Models Tools"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "transformers>=4.44.2",
  "jinja2>=3.1.4",
  "sseclient-py",
  "pytz"
]

[project.urls]
Homepage = "https://github.com/Equiron-AI/smart-context"